+++ src/iter.rs
@@ -6,7 +6,7 @@ use core::mem::ManuallyDrop;
 use core::{cmp, fmt, mem, ptr};
 
 /// An iterator that moves out of a `GenericArray`
-pub struct GenericArrayIter<T, N: ArrayLength<T>> {
+pub struct GenericArrayIter<T, N: ArrayLength> {
     // Invariants: index <= index_back <= N
     // Only values in array[index..index_back] are alive at any given time.
     // Values from array[..index] and array[index_back..] are already moved/dropped.
@@ -27,10 +27,7 @@ mod test {
     }
 }
 
-impl<T, N> GenericArrayIter<T, N>
-where
-    N: ArrayLength<T>,
-{
+impl<T, N: ArrayLength> GenericArrayIter<T, N> {
     /// Returns the remaining items of this iterator as a slice
     #[inline]
     pub fn as_slice(&self) -> &[T] {
@@ -44,10 +41,7 @@ where
     }
 }
 
-impl<T, N> IntoIterator for GenericArray<T, N>
-where
-    N: ArrayLength<T>,
-{
+impl<T, N: ArrayLength> IntoIterator for GenericArray<T, N> {
     type Item = T;
     type IntoIter = GenericArrayIter<T, N>;
 
@@ -61,10 +55,7 @@ where
 }
 
 // Based on work in rust-lang/rust#49000
-impl<T: fmt::Debug, N> fmt::Debug for GenericArrayIter<T, N>
-where
-    N: ArrayLength<T>,
-{
+impl<T: fmt::Debug, N: ArrayLength> fmt::Debug for GenericArrayIter<T, N> {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         f.debug_tuple("GenericArrayIter")
             .field(&self.as_slice())
@@ -72,10 +63,7 @@ where
     }
 }
 
-impl<T, N> Drop for GenericArrayIter<T, N>
-where
-    N: ArrayLength<T>,
-{
+impl<T, N: ArrayLength> Drop for GenericArrayIter<T, N> {
     #[inline]
     fn drop(&mut self) {
         if mem::needs_drop::<T>() {
@@ -90,10 +78,7 @@ where
 }
 
 // Based on work in rust-lang/rust#49000
-impl<T: Clone, N> Clone for GenericArrayIter<T, N>
-where
-    N: ArrayLength<T>,
-{
+impl<T: Clone, N: ArrayLength> Clone for GenericArrayIter<T, N> {
     fn clone(&self) -> Self {
         // This places all cloned elements at the start of the new array iterator,
         // not at their original indices.
@@ -101,7 +86,7 @@ where
         let mut array = unsafe { ptr::read(&self.array) };
         let mut index_back = 0;
 
-        for (dst, src) in array.as_mut_slice().into_iter().zip(self.as_slice()) {
+        for (dst, src) in array.as_mut_slice().iter_mut().zip(self.as_slice()) {
             unsafe { ptr::write(dst, src.clone()) };
             index_back += 1;
         }
@@ -114,10 +99,7 @@ where
     }
 }
 
-impl<T, N> Iterator for GenericArrayIter<T, N>
-where
-    N: ArrayLength<T>,
-{
+impl<T, N: ArrayLength> Iterator for GenericArrayIter<T, N> {
     type Item = T;
 
     #[inline]
@@ -194,10 +176,7 @@ where
     }
 }
 
-impl<T, N> DoubleEndedIterator for GenericArrayIter<T, N>
-where
-    N: ArrayLength<T>,
-{
+impl<T, N: ArrayLength> DoubleEndedIterator for GenericArrayIter<T, N> {
     fn next_back(&mut self) -> Option<T> {
         if self.index < self.index_back {
             self.index_back -= 1;
@@ -237,15 +216,12 @@ where
     }
 }
 
-impl<T, N> ExactSizeIterator for GenericArrayIter<T, N>
-where
-    N: ArrayLength<T>,
-{
+impl<T, N: ArrayLength> ExactSizeIterator for GenericArrayIter<T, N> {
     fn len(&self) -> usize {
         self.index_back - self.index
     }
 }
 
-impl<T, N> FusedIterator for GenericArrayIter<T, N> where N: ArrayLength<T> {}
+impl<T, N: ArrayLength> FusedIterator for GenericArrayIter<T, N> {}
 
 // TODO: Implement `TrustedLen` when stabilized